Spec needs to load rspec helper and the unit under test
If spec_helper existed, we could have a smaller rspec configuration that
did not depend on rails. However, without it, we would need the spec to
load the unit under test as well as rspec and its configuration. It
seems the pattern is to load rails_helper in all specs in order to
configure rspec while also getting rails' autoloading.